+++ src/java/rexos/mas/service_agent/Service.java
@@ -63,7 +63,7 @@ public interface Service {
 	 * @return An array of module id's of the modules needed to perform the
 	 *         service.
 	 */
-	public long[] getModuleIds(long productStepType, BasicDBObject parameters);
+	public int[] getModuleIds(int productStepType, BasicDBObject parameters);
 
 	/**
 	 * This method translates the productStep into a single or multiple
@@ -76,7 +76,7 @@ public interface Service {
 	 * @return An array of ServiceStepMessage each representing a serviceStep as
 	 *         a result of the translation.
 	 */
-	public ServiceStepMessage[] getServiceSteps(long productStepType,
+	public ServiceStepMessage[] getServiceSteps(int productStepType,
 			BasicDBObject parameters);
 
 	/**
@@ -94,7 +94,7 @@ public interface Service {
 	 * 
 	 * @return the id of this service
 	 */
-	public long getId();
+	public int getId();
 
 	/**
 	 * Implementations should return a clear name for this service analog to its
